requirements.txt
flask
requests
bs4
openai
python-dotenv